[Chore] #90 카페 리스트 화면, 필터 하단시트뷰 CafeFilterBottomSheetViewState 관련 파일 분리
